+++ packages/react/src/components/ActionModal/index.ts
@@ -1,0 +1,1 @@
+export { default, ActionModalProps } from './ActionModal';